🎨 Update code style
